+++ Dynamic/Types/FreeTextType.php
@@ -31,7 +31,9 @@ class FreeTextType implements FormFieldTypeInterface
     {
         return new FormFieldTypeConfiguration(
             'sulu_form.type.freetext',
-            'SuluFormBundle:forms:fields/types/freetext.html.twig'
+            'SuluFormBundle:forms:fields/types/freetext.html.twig',
+            [],
+            'additional'
         );
     }