+++ Controller/TemplateController.php
@@ -51,6 +51,8 @@ class TemplateController extends Controller
                 'types' => $this->getSortedTypes($types),
                 'widths' => $widths,
                 'receiverTypes' => $receiverTypes,
+                'locale' => $request->get('locale', $request->getLocale()),
+                'autoTitle' => $this->getParameter('sulu_form.dynamic_auto_title'),
                 'fallbackEmails' => [
                     'from' => $this->getParameter('sulu_form.mail.from'),
                     'to' => $this->getParameter('sulu_form.mail.to'),