+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.4.0 (unreleased)
 
+ - FEATURE     #116    Fixed unkown parameter secret by using kernel.secret parameter
+ - FEATURE     #115    Fixed grid width around submit button
+ - FEATURE     #114    Fixed grid row around submit button
  - FEATURE     #113    Fixed last width when next has no place
  - FEATURE     #109    Add symfony 3 to test runs
  - FEATURE     #108    Add widths to bundle configuration